web/requirements.txt
diff-match-patch==20121119
Django==1.7.4
django-djconfig==0.1.8
django-gravatar2==1.2.1
django-haystack==2.3.1
django-infinite-scroll-pagination==0.1.3
django-markdown-deux==1.0.5
django-model-utils==2.2
django-pagedown==0.1.0
django-separatedvaluesfield==0.2
django-sortedm2m==0.9.5
docutils==0.12
jsmin==2.1.1
mailjet==1.4.0
markdown2==2.3.0
mistune==0.3.1
Pillow==2.6.2
pytz==2014.10
sqlparse==0.1.14
Whoosh==2.5.7
gunicorn
psycopg2
